First attempt at implementing a folder system.
I added a home button, back button, and forward button for easier
navigation. All of the code that was in the initial implementation of the
run function was moved to another function so that it may be used to open
any folder. The structure of the test.txt file was changed as well. The name
of each folder is mapped to a key-value store. I think we could come up
with a better way to manage the folders but this was just my first
idea.